requirements.txt
streamlit
pandas
numpy
Pillow
colormath
matplotlib